Rename network to server graph (finally)
