+++ java/de/uni_kl/cs/discodnc/FF_3S_1SC_2F_1AC_2P_ServerGraph.java
@@ -31,59 +31,59 @@ package de.uni_kl.cs.discodnc;
 import de.uni_kl.cs.discodnc.curves.ArrivalCurve;
 import de.uni_kl.cs.discodnc.curves.Curve;
 import de.uni_kl.cs.discodnc.curves.ServiceCurve;
-import de.uni_kl.cs.discodnc.network.Flow;
-import de.uni_kl.cs.discodnc.network.Network;
-import de.uni_kl.cs.discodnc.network.NetworkFactory;
-import de.uni_kl.cs.discodnc.network.Server;
+import de.uni_kl.cs.discodnc.server_graph.Flow;
+import de.uni_kl.cs.discodnc.server_graph.Link;
+import de.uni_kl.cs.discodnc.server_graph.ServerGraph;
+import de.uni_kl.cs.discodnc.server_graph.ServerGraphFactory;
+import de.uni_kl.cs.discodnc.server_graph.Server;
 
-public class TR_7S_1SC_3F_1AC_3P_Network implements NetworkFactory {
+import java.util.LinkedList;
+
+public class FF_3S_1SC_2F_1AC_2P_ServerGraph implements ServerGraphFactory {
 	private final int sc_R = 20;
 	private final int sc_T = 20;
 	private final int ac_r = 5;
 	private final int ac_b = 25;
 	
-	private Server s0, s1, s2, s3, s4, s5, s6;
+	private Server s0, s1, s2;
+	private Link l_s0_s1, l_s1_s2;
 	
 	private ServiceCurve service_curve = Curve.getFactory().createRateLatency(sc_R, sc_T);
 	private ArrivalCurve arrival_curve = Curve.getFactory().createTokenBucket(ac_r, ac_b);
 	
-	private Network network;
+	private ServerGraph network;
 
-	public TR_7S_1SC_3F_1AC_3P_Network() {
+	public FF_3S_1SC_2F_1AC_2P_ServerGraph() {
 		network = createNetwork();
 	}
 
-	public Network getNetwork() {
+	public ServerGraph getNetwork() {
 		return network;
 	}
 
-	public Network createNetwork() {
-		network = new Network();
+	public ServerGraph createNetwork() {
+		network = new ServerGraph();
 
-		s0 = network.addServer(service_curve);
-		s1 = network.addServer(service_curve);
-		s2 = network.addServer(service_curve);
-		s3 = network.addServer(service_curve);
-		s4 = network.addServer(service_curve);
-		s5 = network.addServer(service_curve);
-		s6 = network.addServer(service_curve);
+		s0 = network.addServer("s0", service_curve);
+		s1 = network.addServer("s1", service_curve);
+		s2 = network.addServer("s2", service_curve);
 
 		try {
-			network.addLink(s0, s1);
-			network.addLink(s1, s2);
-			network.addLink(s2, s5);
-			network.addLink(s3, s4);
-			network.addLink(s4, s5);
-			network.addLink(s5, s6);
+			l_s0_s1 = network.addLink("l_s0_s1", s0, s1);
+			l_s1_s2 = network.addLink("l_s1_s2", s1, s2);
+			network.addLink(s0, s2);
 		} catch (Exception e) {
 			e.printStackTrace();
 			throw new RuntimeException(e);
 		}
 
+		LinkedList<Link> f1_path = new LinkedList<Link>();
+		f1_path.add(l_s0_s1);
+		f1_path.add(l_s1_s2);
+
 		try {
-			network.addFlow("f0", arrival_curve, s1, s6);
-			network.addFlow("f1", arrival_curve, s0, s6);
-			network.addFlow("f2", arrival_curve, s3, s6);
+			network.addFlow("f0", arrival_curve, s0, s2);
+			network.addFlow("f1", arrival_curve, f1_path);
 		} catch (Exception e) {
 			e.printStackTrace();
 			throw new RuntimeException(e);
