+++ java/de/uni_kl/cs/discodnc/FF_4S_1SC_3F_1AC_3P_ServerGraph.java
@@ -31,55 +31,65 @@ package de.uni_kl.cs.discodnc;
 import de.uni_kl.cs.discodnc.curves.ArrivalCurve;
 import de.uni_kl.cs.discodnc.curves.Curve;
 import de.uni_kl.cs.discodnc.curves.ServiceCurve;
-import de.uni_kl.cs.discodnc.network.Flow;
-import de.uni_kl.cs.discodnc.network.Network;
-import de.uni_kl.cs.discodnc.network.NetworkFactory;
-import de.uni_kl.cs.discodnc.network.Server;
+import de.uni_kl.cs.discodnc.server_graph.Flow;
+import de.uni_kl.cs.discodnc.server_graph.Link;
+import de.uni_kl.cs.discodnc.server_graph.ServerGraph;
+import de.uni_kl.cs.discodnc.server_graph.ServerGraphFactory;
+import de.uni_kl.cs.discodnc.server_graph.Server;
 
-public class TA_2S_2SC_1F_1AC_1P_Network implements NetworkFactory {
-	private final int sc_R_0 = 10;
-	private final int sc_T_0 = 10;
-	private final int sc_R_1 = 6;
-	private final int sc_T_1 = 6;
+import java.util.LinkedList;
+import java.util.List;
+
+public class FF_4S_1SC_3F_1AC_3P_ServerGraph implements ServerGraphFactory {
+	private final int sc_R = 20;
+	private final int sc_T = 20;
 	private final int ac_r = 5;
 	private final int ac_b = 25;
 	
-	private Server s0, s1;
+	private Server s0, s1, s2, s3;
+	private Link l_s0_s1, l_s1_s3;
 	
-	private ServiceCurve service_curve_0 = Curve.getFactory().createRateLatency(sc_R_0, sc_T_0);
-	private ServiceCurve service_curve_1 = Curve.getFactory().createRateLatency(sc_R_1, sc_T_1);
+	private ServiceCurve service_curve = Curve.getFactory().createRateLatency(sc_R, sc_T);
 	private ArrivalCurve arrival_curve = Curve.getFactory().createTokenBucket(ac_r, ac_b);
 	
-	private Network network;
+	private ServerGraph network;
 
-	public TA_2S_2SC_1F_1AC_1P_Network() {
+	public FF_4S_1SC_3F_1AC_3P_ServerGraph() {
 		network = createNetwork();
 	}
 
-	public Network getNetwork() {
+	public ServerGraph getNetwork() {
 		return network;
 	}
 
-	public Network createNetwork() {
-		network = new Network();
+	public ServerGraph createNetwork() {
+		network = new ServerGraph();
 
-		s0 = network.addServer(service_curve_0);
-		s0.setUseGamma(false);
-		s0.setUseExtraGamma(false);
-
-		s1 = network.addServer(service_curve_1);
-		s1.setUseGamma(false);
-		s1.setUseExtraGamma(false);
+		s0 = network.addServer(service_curve);
+		s1 = network.addServer(service_curve);
+		s2 = network.addServer(service_curve);
+		s3 = network.addServer(service_curve);
 
 		try {
-			network.addLink(s0, s1);
+			l_s0_s1 = network.addLink(s0, s1);
+			network.addLink(s0, s3);
+			l_s1_s3 = network.addLink(s1, s3);
+			network.addLink(s2, s0);
+			network.addLink(s2, s1);
+			network.addLink(s2, s3);
 		} catch (Exception e) {
 			e.printStackTrace();
 			throw new RuntimeException(e);
 		}
 
+		List<Link> f0_path = new LinkedList<Link>();
+		f0_path.add(l_s0_s1);
+		f0_path.add(l_s1_s3);
+
 		try {
-			network.addFlow("f0", arrival_curve, s0, s1);
+			network.addFlow(arrival_curve, f0_path); // f0
+			network.addFlow(arrival_curve, s2, s3);  // f1
+			network.addFlow(arrival_curve, s2, s1);  // f1
 		} catch (Exception e) {
 			e.printStackTrace();
 			throw new RuntimeException(e);
@@ -89,11 +99,10 @@ public class TA_2S_2SC_1F_1AC_1P_Network implements NetworkFactory {
 	}
 
 	public void reinitializeCurves() {
-		service_curve_0 = Curve.getFactory().createRateLatency(sc_R_0, sc_T_0);
-		s0.setServiceCurve(service_curve_0);
-
-		service_curve_1 = Curve.getFactory().createRateLatency(sc_R_1, sc_T_1);
-		s1.setServiceCurve(service_curve_1);
+		service_curve = Curve.getFactory().createRateLatency(sc_R, sc_T);
+		for (Server server : network.getServers()) {
+			server.setServiceCurve(service_curve);
+		}
 
 		arrival_curve = Curve.getFactory().createTokenBucket(ac_r, ac_b);
 		for (Flow flow : network.getFlows()) {
