+++ java/de/uni_kl/cs/discodnc/S_1SC_10F_10AC_Test.java
@@ -31,15 +31,15 @@ package de.uni_kl.cs.discodnc;
 import de.uni_kl.cs.discodnc.feedforward.analyses.PmooAnalysis;
 import de.uni_kl.cs.discodnc.feedforward.analyses.SeparateFlowAnalysis;
 import de.uni_kl.cs.discodnc.feedforward.analyses.TotalFlowAnalysis;
-import de.uni_kl.cs.discodnc.network.Flow;
-import de.uni_kl.cs.discodnc.network.Network;
+import de.uni_kl.cs.discodnc.server_graph.Flow;
+import de.uni_kl.cs.discodnc.server_graph.ServerGraph;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 
 public class S_1SC_10F_10AC_Test extends DncTest {
-	private static S_1SC_10F_10AC_Network network_factory = new S_1SC_10F_10AC_Network();
-	private static Network network = network_factory.getNetwork();
+	private static S_1SC_10F_10AC_ServerGraph network_factory = new S_1SC_10F_10AC_ServerGraph();
+	private static ServerGraph network = network_factory.getNetwork();
 	private Flow f0, f6;
 
 	private S_1SC_10F_10AC_Test() throws Exception {
