+++ java/de/uni_kl/cs/discodnc/TA_2S_1SC_2F_1AC_1P_ServerGraph.java
@@ -31,12 +31,12 @@ package de.uni_kl.cs.discodnc;
 import de.uni_kl.cs.discodnc.curves.ArrivalCurve;
 import de.uni_kl.cs.discodnc.curves.Curve;
 import de.uni_kl.cs.discodnc.curves.ServiceCurve;
-import de.uni_kl.cs.discodnc.network.Flow;
-import de.uni_kl.cs.discodnc.network.Network;
-import de.uni_kl.cs.discodnc.network.NetworkFactory;
-import de.uni_kl.cs.discodnc.network.Server;
+import de.uni_kl.cs.discodnc.server_graph.Flow;
+import de.uni_kl.cs.discodnc.server_graph.ServerGraph;
+import de.uni_kl.cs.discodnc.server_graph.ServerGraphFactory;
+import de.uni_kl.cs.discodnc.server_graph.Server;
 
-public class TA_2S_1SC_2F_1AC_1P_Network implements NetworkFactory {
+public class TA_2S_1SC_2F_1AC_1P_ServerGraph implements ServerGraphFactory {
 	private final int sc_R = 10;
 	private final int sc_T = 10;
 	private final int ac_r = 5;
@@ -47,18 +47,18 @@ public class TA_2S_1SC_2F_1AC_1P_Network implements NetworkFactory {
 	private ServiceCurve service_curve = Curve.getFactory().createRateLatency(sc_R, sc_T);
 	private ArrivalCurve arrival_curve = Curve.getFactory().createTokenBucket(ac_r, ac_b);
 	
-	private Network network;
+	private ServerGraph network;
 
-	public TA_2S_1SC_2F_1AC_1P_Network() {
+	public TA_2S_1SC_2F_1AC_1P_ServerGraph() {
 		network = createNetwork();
 	}
 
-	public Network getNetwork() {
+	public ServerGraph getNetwork() {
 		return network;
 	}
 
-	public Network createNetwork() {
-		network = new Network();
+	public ServerGraph createNetwork() {
+		network = new ServerGraph();
 
 		s0 = network.addServer(service_curve);
 		s0.setUseGamma(false);
