+++ java/de/uni_kl/cs/discodnc/TA_2S_1SC_4F_1AC_1P_ServerGraph.java
@@ -31,34 +31,34 @@ package de.uni_kl.cs.discodnc;
 import de.uni_kl.cs.discodnc.curves.ArrivalCurve;
 import de.uni_kl.cs.discodnc.curves.Curve;
 import de.uni_kl.cs.discodnc.curves.ServiceCurve;
-import de.uni_kl.cs.discodnc.network.Flow;
-import de.uni_kl.cs.discodnc.network.Network;
-import de.uni_kl.cs.discodnc.network.NetworkFactory;
-import de.uni_kl.cs.discodnc.network.Server;
+import de.uni_kl.cs.discodnc.server_graph.Flow;
+import de.uni_kl.cs.discodnc.server_graph.ServerGraph;
+import de.uni_kl.cs.discodnc.server_graph.ServerGraphFactory;
+import de.uni_kl.cs.discodnc.server_graph.Server;
 
-public class TA_2S_1SC_1F_1AC_1P_Network implements NetworkFactory {
+public class TA_2S_1SC_4F_1AC_1P_ServerGraph implements ServerGraphFactory {
 	private final int sc_R = 10;
 	private final int sc_T = 10;
-	private final int ac_r = 5;
-	private final int ac_b = 25;
+	private final int ac_r = 2;
+	private final int ac_b = 10;
 	
 	private Server s0, s1;
 	
 	private ServiceCurve service_curve = Curve.getFactory().createRateLatency(sc_R, sc_T);
 	private ArrivalCurve arrival_curve = Curve.getFactory().createTokenBucket(ac_r, ac_b);
 	
-	private Network network;
+	private ServerGraph network;
 
-	public TA_2S_1SC_1F_1AC_1P_Network() {
+	public TA_2S_1SC_4F_1AC_1P_ServerGraph() {
 		network = createNetwork();
 	}
 
-	public Network getNetwork() {
+	public ServerGraph getNetwork() {
 		return network;
 	}
 
-	public Network createNetwork() {
-		network = new Network();
+	public ServerGraph createNetwork() {
+		network = new ServerGraph();
 
 		s0 = network.addServer(service_curve);
 		s0.setUseGamma(false);
@@ -77,6 +77,9 @@ public class TA_2S_1SC_1F_1AC_1P_Network implements NetworkFactory {
 
 		try {
 			network.addFlow("f0", arrival_curve, s0, s1);
+			network.addFlow("f1", arrival_curve, s0, s1);
+			network.addFlow("f2", arrival_curve, s0, s1);
+			network.addFlow("f3", arrival_curve, s0, s1);
 		} catch (Exception e) {
 			e.printStackTrace();
 			throw new RuntimeException(e);
