+++ java/de/uni_kl/cs/discodnc/TA_2S_2SC_1F_1AC_1P_ServerGraph.java
@@ -31,59 +31,55 @@ package de.uni_kl.cs.discodnc;
 import de.uni_kl.cs.discodnc.curves.ArrivalCurve;
 import de.uni_kl.cs.discodnc.curves.Curve;
 import de.uni_kl.cs.discodnc.curves.ServiceCurve;
-import de.uni_kl.cs.discodnc.network.Flow;
-import de.uni_kl.cs.discodnc.network.Link;
-import de.uni_kl.cs.discodnc.network.Network;
-import de.uni_kl.cs.discodnc.network.NetworkFactory;
-import de.uni_kl.cs.discodnc.network.Server;
+import de.uni_kl.cs.discodnc.server_graph.Flow;
+import de.uni_kl.cs.discodnc.server_graph.ServerGraph;
+import de.uni_kl.cs.discodnc.server_graph.ServerGraphFactory;
+import de.uni_kl.cs.discodnc.server_graph.Server;
 
-import java.util.LinkedList;
-
-public class FF_3S_1SC_2F_1AC_2P_Network implements NetworkFactory {
-	private final int sc_R = 20;
-	private final int sc_T = 20;
+public class TA_2S_2SC_1F_1AC_1P_ServerGraph implements ServerGraphFactory {
+	private final int sc_R_0 = 10;
+	private final int sc_T_0 = 10;
+	private final int sc_R_1 = 6;
+	private final int sc_T_1 = 6;
 	private final int ac_r = 5;
 	private final int ac_b = 25;
 	
-	private Server s0, s1, s2;
-	private Link l_s0_s1, l_s1_s2;
+	private Server s0, s1;
 	
-	private ServiceCurve service_curve = Curve.getFactory().createRateLatency(sc_R, sc_T);
+	private ServiceCurve service_curve_0 = Curve.getFactory().createRateLatency(sc_R_0, sc_T_0);
+	private ServiceCurve service_curve_1 = Curve.getFactory().createRateLatency(sc_R_1, sc_T_1);
 	private ArrivalCurve arrival_curve = Curve.getFactory().createTokenBucket(ac_r, ac_b);
 	
-	private Network network;
+	private ServerGraph network;
 
-	public FF_3S_1SC_2F_1AC_2P_Network() {
+	public TA_2S_2SC_1F_1AC_1P_ServerGraph() {
 		network = createNetwork();
 	}
 
-	public Network getNetwork() {
+	public ServerGraph getNetwork() {
 		return network;
 	}
 
-	public Network createNetwork() {
-		network = new Network();
+	public ServerGraph createNetwork() {
+		network = new ServerGraph();
 
-		s0 = network.addServer("s0", service_curve);
-		s1 = network.addServer("s1", service_curve);
-		s2 = network.addServer("s2", service_curve);
+		s0 = network.addServer(service_curve_0);
+		s0.setUseGamma(false);
+		s0.setUseExtraGamma(false);
+
+		s1 = network.addServer(service_curve_1);
+		s1.setUseGamma(false);
+		s1.setUseExtraGamma(false);
 
 		try {
-			l_s0_s1 = network.addLink("l_s0_s1", s0, s1);
-			l_s1_s2 = network.addLink("l_s1_s2", s1, s2);
-			network.addLink(s0, s2);
+			network.addLink(s0, s1);
 		} catch (Exception e) {
 			e.printStackTrace();
 			throw new RuntimeException(e);
 		}
 
-		LinkedList<Link> f1_path = new LinkedList<Link>();
-		f1_path.add(l_s0_s1);
-		f1_path.add(l_s1_s2);
-
 		try {
-			network.addFlow("f0", arrival_curve, s0, s2);
-			network.addFlow("f1", arrival_curve, f1_path);
+			network.addFlow("f0", arrival_curve, s0, s1);
 		} catch (Exception e) {
 			e.printStackTrace();
 			throw new RuntimeException(e);
@@ -93,10 +89,11 @@ public class FF_3S_1SC_2F_1AC_2P_Network implements NetworkFactory {
 	}
 
 	public void reinitializeCurves() {
-		service_curve = Curve.getFactory().createRateLatency(sc_R, sc_T);
-		for (Server server : network.getServers()) {
-			server.setServiceCurve(service_curve);
-		}
+		service_curve_0 = Curve.getFactory().createRateLatency(sc_R_0, sc_T_0);
+		s0.setServiceCurve(service_curve_0);
+
+		service_curve_1 = Curve.getFactory().createRateLatency(sc_R_1, sc_T_1);
+		s1.setServiceCurve(service_curve_1);
 
 		arrival_curve = Curve.getFactory().createTokenBucket(ac_r, ac_b);
 		for (Flow flow : network.getFlows()) {
