+++ java/de/uni_kl/cs/discodnc/TA_3S_1SC_2F_1AC_1P_ServerGraph.java
@@ -31,43 +31,50 @@ package de.uni_kl.cs.discodnc;
 import de.uni_kl.cs.discodnc.curves.ArrivalCurve;
 import de.uni_kl.cs.discodnc.curves.Curve;
 import de.uni_kl.cs.discodnc.curves.ServiceCurve;
-import de.uni_kl.cs.discodnc.network.Flow;
-import de.uni_kl.cs.discodnc.network.Network;
-import de.uni_kl.cs.discodnc.network.NetworkFactory;
-import de.uni_kl.cs.discodnc.network.Server;
+import de.uni_kl.cs.discodnc.server_graph.Flow;
+import de.uni_kl.cs.discodnc.server_graph.ServerGraph;
+import de.uni_kl.cs.discodnc.server_graph.ServerGraphFactory;
+import de.uni_kl.cs.discodnc.server_graph.Server;
 
-public class S_1SC_10F_10AC_Network implements NetworkFactory {
-	private final int sc_R = 10;
-	private final int sc_T = 10;
+public class TA_3S_1SC_2F_1AC_1P_ServerGraph implements ServerGraphFactory {
+	private final int sc_R = 20;
+	private final int sc_T = 20;
+	private final int ac_r = 5;
+	private final int ac_b = 25;
 	
-	private Server s0;
-	private Flow[] flows = new Flow[10];
+	protected Server s0, s1, s2;
 	
 	private ServiceCurve service_curve = Curve.getFactory().createRateLatency(sc_R, sc_T);
-	private ArrivalCurve[] arrival_curves = new ArrivalCurve[10];
+	private ArrivalCurve arrival_curve = Curve.getFactory().createTokenBucket(ac_r, ac_b);
 	
-	private Network network;
+	private ServerGraph network;
 
-	public S_1SC_10F_10AC_Network() {
+	public TA_3S_1SC_2F_1AC_1P_ServerGraph() {
 		network = createNetwork();
 	}
 
-	public Network getNetwork() {
+	public ServerGraph getNetwork() {
 		return network;
 	}
 
-	public Network createNetwork() {
-		network = new Network();
+	public ServerGraph createNetwork() {
+		network = new ServerGraph();
 
 		s0 = network.addServer(service_curve);
-		s0.setUseGamma(false);
-		s0.setUseExtraGamma(false);
+		s1 = network.addServer(service_curve);
+		s2 = network.addServer(service_curve);
 
 		try {
-			for (int i = 1; i <= 10; i++) {
-				arrival_curves[i - 1] = Curve.getFactory().createTokenBucket(i * 0.1, i);
-				flows[i - 1] = network.addFlow( "f" + Integer.toString(i-1) , arrival_curves[i - 1], s0);
-			}
+			network.addLink(s0, s1);
+			network.addLink(s1, s2);
+		} catch (Exception e) {
+			e.printStackTrace();
+			throw new RuntimeException(e);
+		}
+
+		try {
+			network.addFlow("f0", arrival_curve, s0, s2);
+			network.addFlow("f1", arrival_curve, s0, s2);
 		} catch (Exception e) {
 			e.printStackTrace();
 			throw new RuntimeException(e);
@@ -82,9 +89,9 @@ public class S_1SC_10F_10AC_Network implements NetworkFactory {
 			server.setServiceCurve(service_curve);
 		}
 
-		for (int i = 1; i <= 10; i++) {
-			arrival_curves[i - 1] = Curve.getFactory().createTokenBucket(i * 0.1, i);
-			flows[i - 1].setArrivalCurve(arrival_curves[i - 1]);
+		arrival_curve = Curve.getFactory().createTokenBucket(ac_r, ac_b);
+		for (Flow flow : network.getFlows()) {
+			flow.setArrivalCurve(arrival_curve);
 		}
 	}
 }
