+++ java/de/uni_kl/cs/discodnc/TA_4S_1SC_2F_1AC_2P_ServerGraph.java
@@ -31,48 +31,52 @@ package de.uni_kl.cs.discodnc;
 import de.uni_kl.cs.discodnc.curves.ArrivalCurve;
 import de.uni_kl.cs.discodnc.curves.Curve;
 import de.uni_kl.cs.discodnc.curves.ServiceCurve;
-import de.uni_kl.cs.discodnc.network.Flow;
-import de.uni_kl.cs.discodnc.network.Network;
-import de.uni_kl.cs.discodnc.network.NetworkFactory;
-import de.uni_kl.cs.discodnc.network.Server;
+import de.uni_kl.cs.discodnc.server_graph.Flow;
+import de.uni_kl.cs.discodnc.server_graph.ServerGraph;
+import de.uni_kl.cs.discodnc.server_graph.ServerGraphFactory;
+import de.uni_kl.cs.discodnc.server_graph.Server;
 
-public class TA_2S_1SC_2F_1AC_2P_Network implements NetworkFactory {
+public class TA_4S_1SC_2F_1AC_2P_ServerGraph implements ServerGraphFactory {
 	private final int sc_R = 20;
 	private final int sc_T = 20;
 	private final int ac_r = 5;
 	private final int ac_b = 25;
 	
-	private Server s0, s1;
+	private Server s0, s1, s2, s3;
 	
 	private ServiceCurve service_curve = Curve.getFactory().createRateLatency(sc_R, sc_T);
 	private ArrivalCurve arrival_curve = Curve.getFactory().createTokenBucket(ac_r, ac_b);
 	
-	private Network network;
+	private ServerGraph network;
 
-	public TA_2S_1SC_2F_1AC_2P_Network() {
+	public TA_4S_1SC_2F_1AC_2P_ServerGraph() {
 		network = createNetwork();
 	}
 
-	public Network getNetwork() {
+	public ServerGraph getNetwork() {
 		return network;
 	}
 
-	public Network createNetwork() {
-		network = new Network();
+	public ServerGraph createNetwork() {
+		network = new ServerGraph();
 
 		s0 = network.addServer(service_curve);
 		s1 = network.addServer(service_curve);
+		s2 = network.addServer(service_curve);
+		s3 = network.addServer(service_curve);
 
 		try {
 			network.addLink(s0, s1);
+			network.addLink(s1, s2);
+			network.addLink(s2, s3);
 		} catch (Exception e) {
 			e.printStackTrace();
 			throw new RuntimeException(e);
 		}
 
 		try {
-			network.addFlow("f0", arrival_curve, s0, s1);
-			network.addFlow("f1", arrival_curve, s1);
+			network.addFlow("f0", arrival_curve, s0, s3);
+			network.addFlow("f1", arrival_curve, s1, s2);
 		} catch (Exception e) {
 			e.printStackTrace();
 			throw new RuntimeException(e);
